refactor: extract use-contextual-menu-actions.ts
